Keep the search for a unit in a module of its own
quantity.ts said both what a quantity is and how to go looking for the units to
write one in, and the second was most of the file. The search moves to
unit-search.ts, which takes the dimensions and hands back a representation.

It imports the vocabulary it works in as types, so nothing runs across the
boundary in that direction. The two pieces it did share code with move to where
they belonged anyway: how a number is written is text.ts's business, and a run
of plain text is human-readable-name's.
